ai/requirements.txt
diskcache==5.6.3
Jinja2==3.1.3
llama_cpp_python==0.2.68
MarkupSafe==2.1.5
numpy==1.26.4
typing_extensions==4.11.0
